repositories: Use mirrors.edge.kernel.org mirror for the linuxmint.com provider
This commit replaces the official mirror for `linuxmint.com` (http://packages.linuxmint.com)
with https://mirrors.edge.kernel.org/linuxmint-packages due to the official one being unreachable sporadically.
This hinders development, testing, and last but not least the usability of `xdeb-install` as a whole.

The mirror will be changed back to the official one once it's fixed, if at all.

Signed-off-by: Timo Reichl <thetredev@gmail.com>